MongoDB Data Transfer
Reordered the (name, value, unit) list of attributes that is sent to save_db function in LogSensors.py to be consistent with Google Sheet. Previously, the "name" variable had importing into the "value" field, rather than the "attribute" field where it belonged.

Also modified RemoteMongoDB.py file to separate date format function for string, used by Howard Webb for logging, and the one which should be sent to MongoDB. This update will now send the observation_date as a datetime object to MongoDB, which will carry with it information about timezone.